correct sign conversion fixes to buffer.c and doc.c
Not correct after all since the indentation can be negative.

The following commits are reverted:

e2d09c2 fix implicit sign conversions in buffer.c
9c2fdf6 fix implicit sign conversions in doc_exec_minimize_indent()
9cb503d fix implicit sign conversions in doc_exec_align()
155b497 ignore implicit sign conversions in doc_exec_indent()
07d0638 fix implicit sign conversions in doc.c

+++ buffer.c
@@ -150,7 +150,7 @@ buffer_vprintf(struct buffer *bf, const char *fmt, va_list ap)
 }
 
 size_t
-buffer_indent(struct buffer *bf, size_t indent, int usetabs, size_t pos)
+buffer_indent(struct buffer *bf, int indent, int usetabs, size_t pos)
 {
 	if (usetabs) {
 		for (; indent >= 8; indent -= 8) {
